修复 vm中的类型转换错误

+++ accessor/LinearValueMapper.cpp
@@ -9,7 +9,7 @@ namespace wibot::accessor
 
 	float LinearValueMapper::value_get(uint32_t raw_value)
 	{
-		return (float)(raw_value - config.zero_offset) * config.value_per_unit;
+		return (float)(int32_t)(raw_value - config.zero_offset) * config.value_per_unit;
 	}
 
 } // wibot::accessor
